fix(backend): make user id bigint

+++ backend/src/database/models.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import Column, DateTime, func, VARCHAR, Integer, ForeignKey
+from sqlalchemy import Column, DateTime, func, VARCHAR, Integer, ForeignKey, BigInteger
 from sqlalchemy.dialects.postgresql import UUID, INTERVAL, ARRAY
 from sqlalchemy.ext.mutable import Mutable
 from sqlalchemy.orm import relationship
@@ -11,7 +11,7 @@ from database.database import Base
 class User(Base):
     __tablename__ = 'user'
 
-    id = Column(Integer, primary_key=True, unique=True, nullable=False)
+    id = Column(BigInteger, primary_key=True, unique=True, nullable=False)
     name = Column(VARCHAR(255), nullable=False)
     timezone = Column(Integer, nullable=False)
     notification_time = Column(Mutable.as_mutable(ARRAY(Integer)), nullable=False)
@@ -26,7 +26,7 @@ class Schedule(Base):
     __tablename__ = 'schedule'
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4, unique=True, nullable=False)
-    user_id = Column(Integer, ForeignKey(User.id), nullable=False)
+    user_id = Column(BigInteger, ForeignKey(User.id), nullable=False)
     windows = Column(Mutable.as_mutable(ARRAY(Integer)), nullable=False)
 
     user = relationship(User, uselist=False)
@@ -36,8 +36,8 @@ class CalendarEvent(Base):
     __tablename__ = 'calendar_event'
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4, unique=True, nullable=False)
-    owner_user_id = Column(Integer, ForeignKey(User.id), nullable=False)
-    invited_user_id = Column(Integer, ForeignKey(User.id), nullable=False)
+    owner_user_id = Column(BigInteger, ForeignKey(User.id), nullable=False)
+    invited_user_id = Column(BigInteger, ForeignKey(User.id), nullable=False)
     appointment_time = Column(DateTime, nullable=False)
     duration = Column(INTERVAL, nullable=False)
     created_at = Column(DateTime, nullable=False, server_default=func.now())
